feat: 더보기, 접기 버튼 훅 분리
