+++ src/App.tsx
@@ -1,9 +1,11 @@
+import TempleInfo from '@components/templeDetail/templeInfo/templeInfo';
 import { Outlet } from 'react-router-dom';
 
 const App = () => {
   return (
     <main style={{ margin: '0 auto' }}>
       <Outlet />
+      <TempleInfo />
     </main>
   );
 };
